+++ src/Text/Pandoc/Filter/Plot/Renderers/GNUPlot.hs
@@ -44,7 +44,7 @@ gnuplot = do
               }
 
 gnuplotSupportedSaveFormats :: [SaveFormat]
-gnuplotSupportedSaveFormats = [PNG, SVG, EPS, GIF, JPG, PDF]
+gnuplotSupportedSaveFormats = [LaTeX, PNG, SVG, EPS, GIF, JPG, PDF]
 
 gnuplotCommand :: Text -> Text -> OutputSpec -> Text
 gnuplotCommand cmdargs exe OutputSpec {..} = [st|#{exe} #{cmdargs} -c "#{oScriptPath}"|]
@@ -66,8 +66,13 @@ gnuplotCaptureFragment :: FigureSpec -> FilePath -> Script
 gnuplotCaptureFragment FigureSpec {..} fname =
   [st|
 set terminal #{terminalString saveFormat}
-set output '#{fname}'
+set output '#{normalizePath fname}'
 |]
+  where
+    normalizePath = map f
+      where
+        f '\\' = '/'
+        f x = x
 
 -- | Terminal name for supported save formats
 terminalString :: SaveFormat -> Text
@@ -77,4 +82,5 @@ terminalString EPS = "postscript eps"
 terminalString GIF = "gif"
 terminalString JPG = "jpeg"
 terminalString PDF = "pdfcairo"
+terminalString LaTeX = "cairolatex"
 terminalString fmt = errorWithoutStackTrace $ "gnuplot: unsupported save format" <> show fmt